+++ src/io/github/mathiasberwig/Utils.java
@@ -7,11 +7,11 @@ public class Utils {
      * @param matrix input matrix.
      * @return transposed with signal-inverted values matrix.
      */
-    public static int[][] trasposeAndChangeSignMatrix(int[][] matrix) {
+    public static double[][] trasposeAndChangeSignMatrix(double[][] matrix) {
         int m = matrix.length;
         int n = matrix[0].length;
 
-        int[][] trasposedMatrix = new int[n][m];
+        double[][] trasposedMatrix = new double[n][m];
 
         for (int x = 0; x < n; x++) {
             for (int y = 0; y < m; y++) {
